Updated to pull WorkItem/WorkItemData out to seperate class so that it can be used for both v1 and v2
